pyproject.toml
[project]
name = "massachusetts-roads-segmentation"
version = "0.1.0"
description = "Semantic segmentation of roads on aerial imagery"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "segmentation-models-pytorch>=0.3.0",
    "albumentations>=1.3.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.13.0",
    "kagglehub>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.3.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "ipykernel>=6.25.0",
]